Add files via upload
you will find 2 seperate python files here , they both does the same thing in different way , these scripts opens a browser in the background and builds a bridge between openchatai and the irc channel , it read from the web pages <p> <li> and <code class> elements  and brings them to the irc channel that is defined in the script , also the nick is defined in the script you can change to anything and the password and the server configuration can all be done within the script , one ohese files merahabaupdatedherkesin.py gets a nick to itself connect to a server and a channel then starts listening every even in the channel and in case of someone mentions his nick in a format like timer:can you write me a code? , with the nick and : next to it will activate the bot in the channel , timer:what is your name? , timer:who are you? etc... for the second file which is merhabaupdated2.py does the same thing but instead of processing command from everyone in the channel this bot just accepts commands from the 5 other nicks that are in the channel that is also defined in the python file ,  you can see an example how this bot works on liberachat irc server #rebel channel , feel free to use but always use with respect and good purpose please , this bot gives great functinalities and horinozns to the topics that are being discussed in the channel , also bot brings  title and description of any link that has been printed in the channel.